services/data-analyzer/src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.134.0",
    "pika>=1.3.2",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psycopg2>=2.9.11",
    "python-dateutil>=2.9.0.post0",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
